+++ app/src/main/java/com/gojol/notto/di/TodoLabelDatabaseModule.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import androidx.room.Room
 import com.gojol.notto.model.database.TodoLabelDatabase
 import com.gojol.notto.model.database.todolabel.TodoLabelDao
-import com.gojol.notto.model.datasource.todo.TodoLabelDataSource
 import com.gojol.notto.model.datasource.todo.TodoLabelLocalDataSource
+import com.gojol.notto.model.datasource.todo.TodoLabelLocalDataSourceImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -33,7 +33,7 @@ class TodoLabelDatabaseModule {
     }
 
     @Provides
-    fun provideLocalDataSource(todoLabelDao: TodoLabelDao): TodoLabelDataSource {
-        return TodoLabelLocalDataSource(todoLabelDao)
+    fun provideTodoLabelLocalDataSource(todoLabelDao: TodoLabelDao): TodoLabelLocalDataSource {
+        return TodoLabelLocalDataSourceImpl(todoLabelDao)
     }
 }
